getrlimit/getrlimit03: Skip test if __NR_getrlimit_ulong not implemented
In the LTP code, __NR_getrlimit_ulong is defined as __NR_ugetrlimit
or __NR_getrlimit.

In the Linux kernel, LoongArch uses the generic syscall table
which is defined in include/uapi/asm-generic/unistd.h. We can
see that __NR_ugetrlimit is not defined in the generic header,
and also there is no __ARCH_WANT_SET_GET_RLIMIT definition in
the arch specific header arch/loongarch/include/asm/unistd.h,
so both __NR_ugetrlimit and __NR_getrlimit are not implemented
on LoongArch.

That is to say, no need to compare the return value and errno
about the syscall numbers __NR_prlimit64 and __NR_{u}getrlimit,
just check this case and then skip the test.

While at it, modify the value of test.tcnt as 1 to skip only
once if the errno of getrlimit_ulong() or getrlimit_long() is
ENOSYS.

Without this patch:

  $ /opt/ltp/testcases/bin/getrlimit03
  ...
  getrlimit03.c:102: TFAIL: __NR_prlimit64(0) returned 0 (SUCCESS) but __NR_getrlimit(0) returned -1 (ENOSYS)
  getrlimit03.c:102: TFAIL: __NR_prlimit64(1) returned 0 (SUCCESS) but __NR_getrlimit(1) returned -1 (ENOSYS)
  getrlimit03.c:102: TFAIL: __NR_prlimit64(2) returned 0 (SUCCESS) but __NR_getrlimit(2) returned -1 (ENOSYS)
  ...
  getrlimit03.c:102: TFAIL: __NR_prlimit64(12) returned 0 (SUCCESS) but __NR_getrlimit(12) returned -1 (ENOSYS)
  getrlimit03.c:102: TFAIL: __NR_prlimit64(13) returned 0 (SUCCESS) but __NR_getrlimit(13) returned -1 (ENOSYS)
  getrlimit03.c:102: TFAIL: __NR_prlimit64(14) returned 0 (SUCCESS) but __NR_getrlimit(14) returned -1 (ENOSYS)
  getrlimit03.c:102: TFAIL: __NR_prlimit64(15) returned 0 (SUCCESS) but __NR_getrlimit(15) returned -1 (ENOSYS)

With this patch:

  $ /opt/ltp/testcases/bin/getrlimit03
  ...
  getrlimit03.c:166: TCONF: __NR_getrlimit not implemented: ENOSYS (38)

Signed-off-by: Tiezhu Yang <yangtiezhu@loongson.cn>
Reviewed-by: Li Wang <liwang@redhat.com>